Detail has joined the fray
Detail is now full functional with the exception of the suitability list.

Detail has its own script, css and page. Moved elements from details over to detail, then deleted details.

The only thing missing is the suitability index, but that's more of a cherry on top than a necessity (still doing it later tho lol)